nk3: Extract basic functionality into trussed module
This patch introduces a pynitrokey.trussed module that contains the
functionality that is common for all Nitrokey devices using the Trussed
framework:  communication via CTAPHID, basic commands using admin-app
and provisioner-app, firmware updates using the Nitrokey container
format and the LPC55 and NRF52 bootloader implementations.
