+++ pynitrokey/trussed/base.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright 2021 Nitrokey Developers
+# Copyright 2021-2024 Nitrokey Developers
 #
 # Licensed under the Apache License, Version 2.0, <LICENSE-APACHE or
 # http://apache.org/licenses/LICENSE-2.0> or the MIT license <LICENSE-MIT or
@@ -10,19 +10,39 @@
 from abc import ABC, abstractmethod
 from typing import Optional, TypeVar
 
+from . import VID_NITROKEY
 from .utils import Uuid
 
-T = TypeVar("T", bound="Nitrokey3Base")
+T = TypeVar("T", bound="NitrokeyTrussedBase")
 
 
-class Nitrokey3Base(ABC):
-    """Base class for Nitrokey 3 devices, running the firmware or the bootloader."""
+class NitrokeyTrussedBase(ABC):
+    """
+    Base class for Nitrokey devices using the Trussed framework and running
+    the firmware or the bootloader.
+    """
 
     def __enter__(self: T) -> T:
         return self
 
     def __exit__(self, exc_type: None, exc_val: None, exc_tb: None) -> None:
         self.close()
+
+    def validate_vid_pid(self, vid: int, pid: int) -> None:
+        if (vid, pid) != (self.vid, self.pid):
+            raise ValueError(
+                f"Not a {self.name} device: expected VID:PID "
+                f"{self.vid:x}:{self.pid:x}, got {vid:x}:{pid:x}"
+            )
+
+    @property
+    def vid(self) -> int:
+        return VID_NITROKEY
+
+    @property
+    @abstractmethod
+    def pid(self) -> int:
+        ...
 
     @property
     @abstractmethod
