+++ pynitrokey/trussed/bootloader/nrf52_upload/README.md
