+++ pynitrokey/trussed/bootloader/nrf52_upload/dfu/crc16.py
