+++ pynitrokey/trussed/bootloader/nrf52_upload/dfu/dfu_transport.py
