+++ pynitrokey/trussed/bootloader/nrf52_upload/dfu/init_packet_pb.py
