+++ pynitrokey/trussed/bootloader/nrf52_upload/dfu/manifest.py
