+++ pynitrokey/trussed/bootloader/nrf52_upload/dfu/model.py
