+++ pynitrokey/trussed/bootloader/nrf52_upload/dfu/nrfhex.py
