+++ pynitrokey/trussed/bootloader/nrf52_upload/dfu/signing.py
