+++ pynitrokey/trussed/bootloader/nrf52_upload/exceptions.py
