+++ pynitrokey/trussed/bootloader/nrf52_upload/lister/device_lister.py
