+++ pynitrokey/trussed/bootloader/nrf52_upload/lister/enumerated_device.py
