+++ pynitrokey/trussed/bootloader/nrf52_upload/lister/windows/constants.py
