+++ pynitrokey/trussed/bootloader/nrf52_upload/lister/windows/lister_win32.py
